- Conf fixes

+++ libmodmqttsrv/config.cpp
@@ -54,8 +54,8 @@ ModbusNetworkConfig::ModbusNetworkConfig(const YAML::Node& source) {
         mDevice = ConfigTools::readRequiredString(source, "device");
         mBaud = ConfigTools::readRequiredValue<int>(source, "baud");
         mParity = ConfigTools::readRequiredValue<char>(source, "parity");
-        mDataBit = ConfigTools::readRequiredValue<uint8_t>(source, "data_bit");
-        mStopBit = ConfigTools::readRequiredValue<uint8_t>(source, "stop_bit");
+        mDataBit = ConfigTools::readRequiredValue<int>(source, "data_bit");
+        mStopBit = ConfigTools::readRequiredValue<int>(source, "stop_bit");
         ConfigTools::readOptionalValue<RtuSerialMode>(this->mRtuSerialMode, source, "rtu_serial_mode");
         ConfigTools::readOptionalValue<RtuRtsMode>(this->mRtsMode, source, "rtu_rts_mode");
         ConfigTools::readOptionalValue<int>(this->mRtsDelayUs, source, "rtu_rts_delay");
